Renames auto-gen classes
